+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/message/ChatEvents/MigratedToSupergroup.kt
@@ -3,6 +3,9 @@ package dev.inmo.tgbotapi.types.message.ChatEvents
 import dev.inmo.tgbotapi.types.ChatId
 import dev.inmo.tgbotapi.types.message.ChatEvents.abstracts.SupergroupEvent
 
-class SupergroupChatCreated(
-    val migratedFrom: ChatId?
+/**
+ * This event is sent when a group is converted to a supergroup.
+ */
+class MigratedToSupergroup(
+    val migratedFrom: ChatId
 ): SupergroupEvent